FreeBSD: drops `bsd-kvm-sys` dep
Fixes small bugs

+++ src/columns/vm_size.rs
@@ -68,7 +68,9 @@ impl Column for VmSize {
 #[cfg(target_os = "freebsd")]
 impl Column for VmSize {
     fn add(&mut self, proc: &ProcessInfo) {
-        let raw_content = proc.curr_proc.ki_size;
+        // `ki_size` is a `vm_size_t`, which is a `usize` in `libc`'s binding
+        // and only 32 bits wide on the 32-bit architectures.
+        let raw_content = proc.curr_proc.ki_size as u64;
         let fmt_content = bytify(raw_content);
 
         self.fmt_contents.insert(proc.pid, fmt_content);
